Refactor, dynamic cast (qgraphicsitem_cast)
Code for shortcuts has been refactored; now in stead of direct
transformations, a simple function is called - short and reusable.
This was changed because of buttons in TemplateWindow. When they
are clicked, aforementioned function for furniture transformation is called,
in stead of writing a new one.
Similar thing happens again in TemplateWindow - when using selectedItems(),
return type is QGraphicsItem*, but qgraphicsitem_cast changes it to Furniture*.